# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem manajemen inventory dengan authentication JWT",
        "title": "Inventory Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/peminjaman": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil semua data peminjaman",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Peminjaman"
                ],
                "summary": "Get all peminjaman",
                "responses": {
                    "200": {
                        "description": "List semua peminjaman",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Peminjaman"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Membuat data peminjaman baru",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Peminjaman"
                ],
                "summary": "Create new peminjaman",
                "parameters": [
                    {
                        "description": "Data peminjaman baru",
                        "name": "peminjaman",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Peminjaman"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Peminjaman berhasil dibuat",
                        "schema": {
                            "$ref": "#/definitions/models.Peminjaman"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Barang tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/peminjaman/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil data peminjaman berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Peminjaman"
                ],
                "summary": "Get peminjaman by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Peminjaman ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data peminjaman",
                        "schema": {
                            "$ref": "#/definitions/models.Peminjaman"
                        }
                    },
                    "400": {
                        "description": "ID tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Data peminjaman tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus data peminjaman berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Peminjaman"
                ],
                "summary": "Delete peminjaman",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Peminjaman ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data peminjaman berhasil dihapus",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "ID tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Data peminjaman tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/peminjaman/{id}/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengupdate status peminjaman (dipinjam/dikembalikan)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Peminjaman"
                ],
                "summary": "Update status peminjaman",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Peminjaman ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status baru",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status berhasil diperbarui",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Peminjaman": {
            "type": "object",
            "properties": {
                "barang_id": {
                    "type": "string"
                },
                "email_peminjam": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jumlah": {
                    "type": "integer"
                },
                "nama_peminjam": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tanggal_pinjam": {
                    "type": "string"
                },
                "telepon_peminjam": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}